Move socket io client to own directory
