+++ src/services/socketIoClient/ISocketIoClient.ts
